chore(package): write es modules to esm instead of es6 directory
BREAKING CHANGE:

Previously, es modules were published to the es6/ directory inside the npm
package. This change publishes them to the esm directory to be consistent
with other angular packages. Currently, the es6 directory is still published
in the npm package, but will be removed in a future release.

Fixes #310