docs(3-retrieving-data-as-lists.md): example clarity
Update component methods to be prefixed by 'Item'. This prevents collisions, and improves readability, in situations where the previous step's example is combined with this one.
Update field being updated from size to text. This allows the updates to be immediately observable without any additional code, since item.size was not being displayed.