Update to Retrieve Data as Object
Trying to make the tutorial consistent for each step. As of now the "Install & setup" Page and "Retrieve Data as Object" page have lot of inconsistencies which doesn't make the tutorial go in flow. Also, the "moduleId" is causing errors and is not available in default "app.component.ts", so removing that attribute. Will clean up other pages, may be tomorrow.