Add reference to #1158 'Proposal'
I wish I had read this prior to attempting to migrate.  It may just be that now that I have stumbled through the migration changes the "Proposal" makes good sense but it seems like a resource that would help others.  Much of the added goodness provided by these API changes need to be more fully explained and I think the proposal does a very good job in doing that, plus the comments section make a very good addition.  

I especially appreciated David's opening remark  "Breaking changes are not taken lightly. If we're going to change something, we're going to make it worth your while. Trust me, these new features are well worth it."  However with only the above migration doc to follow, I was cursing some unnamed person or persons long and hard over the past several days.   

I consider myself a skilled novice in all this RxJS-NgRx, like the majority of the folks who would be attempting this migration, and my feedback is that this was far from trivial search and replace.  The end result was much cleaner and I can say that you all have delivered on David's promise, but any additional documentation to assist is going to be very much appreciated.