fix(auth-router) Broken canActivate() helper logic
Fixes #2099 

`canActivate()` has the ternary of `pipe.name === '' ? pipe : () => pipe` which seems to be intended to tell if the `pipe` is in either this format `() => pipe` or `pipe` format currently. If it in `pipe` format then the output of `pipe.name` seems to be `piped` however if it is  in `() => pipe` format the the result of `pipe.name` is actually the name of the function which appears to what has been overlooked with the original logic as `pipe.name` never seems to be an empty string.

Adjusting to `pipe.name !== 'piped' ? pipe : () => pipe` should fix this issue in all situations unless the name of the function is `piped` which would result in `authGuardPipe` being set to `() => () => pipe` which in my testing seems to behave has expected still activating the pipe however I am unsure of the exact ramifications of this.

I have only been playing with this for about an hour or two, so if there is something I am missing definitely let me know.