Add TDD to the instructions
It feels a little unfair to leave the user totally on their own to create the tests. Mocking in TypeScript and JavaScript is confusing for beginners (like me), and it's vital to creating a maintainable codebase. This adds a very quickly put together an example of how to develop this example in a TDD style (while hopefully not being distracting).

This is important because quickstarts become the applications of the future, and a bad first impression in testing Angular and Firebase might put them off doing it.