fix(deploy): remove direct workspace access
Depending on the configuration and initialization of the builder runtime, there may not be a workspace file or the raw values within the workspace file may not represent the final values used by a builder.  To support cases where inspection of builder target options is needed, the builder runtime provides a utility function (`getTargetOptions`) within the context object passed to each builder.  This utility function is now used within the deploy builder to access builder target options.  Also, as a result of these changes, usage of the experimental workspace API was removed.  This experimental API is no longer present as of Angular v11.