fix: Avoid handling promise rejections twice in stability helper (#3657)
`PendingTasks.run` reports rejections of the promise to the
`ErrorHandler`. This function already returns a promise that is meant to
be handled in the developer's application. If there _is_ handling there,
the rejection is handled both in the `PendingTasks.run` and in the
developer callsite.

This change updates the code to use `PendingTasks.add` instead, which
has no built in error handling mechanisms.

fixes https://github.com/angular/angular/issues/61932